Refactoring NotificationPreferencesView
