+++ app/views/NotificationPreferencesView/index.js
@@ -17,126 +17,10 @@ import { withTheme } from '../../theme';
 import protectedFunction from '../../lib/methods/helpers/protectedFunction';
 import SafeAreaView from '../../containers/SafeAreaView';
 import log, { events, logEvent } from '../../utils/log';
-
-const SectionTitle = React.memo(({ title, theme }) => (
-	<Text
-		style={[
-			styles.sectionTitle,
-			{
-				backgroundColor: themes[theme].auxiliaryBackground,
-				color: themes[theme].infoText
-			}
-		]}
-	>
-		{title}
-	</Text>
-));
-
-const SectionSeparator = React.memo(({ theme }) => (
-	<View
-		style={[
-			styles.sectionSeparatorBorder,
-			{ backgroundColor: themes[theme].auxiliaryBackground }
-		]}
-	/>
-));
-
-const Info = React.memo(({ info, theme }) => (
-	<Text
-		style={[
-			styles.infoText,
-			{
-				color: themes[theme].infoText,
-				backgroundColor: themes[theme].auxiliaryBackground
-			}
-		]}
-	>
-		{info}
-	</Text>
-));
-
-SectionTitle.propTypes = {
-	title: PropTypes.string,
-	theme: PropTypes.string
-};
-
-SectionSeparator.propTypes = {
-	theme: PropTypes.string
-};
-
-Info.propTypes = {
-	info: PropTypes.string,
-	theme: PropTypes.string
-};
-
-const OPTIONS = {
-	desktopNotifications: [{
-		label: 'Default', value: 'default'
-	}, {
-		label: 'All_Messages', value: 'all'
-	}, {
-		label: 'Mentions', value: 'mentions'
-	}, {
-		label: 'Nothing', value: 'nothing'
-	}],
-	audioNotifications: [{
-		label: 'Default', value: 'default'
-	}, {
-		label: 'All_Messages', value: 'all'
-	}, {
-		label: 'Mentions', value: 'mentions'
-	}, {
-		label: 'Nothing', value: 'nothing'
-	}],
-	mobilePushNotifications: [{
-		label: 'Default', value: 'default'
-	}, {
-		label: 'All_Messages', value: 'all'
-	}, {
-		label: 'Mentions', value: 'mentions'
-	}, {
-		label: 'Nothing', value: 'nothing'
-	}],
-	emailNotifications: [{
-		label: 'Default', value: 'default'
-	}, {
-		label: 'All_Messages', value: 'all'
-	}, {
-		label: 'Mentions', value: 'mentions'
-	}, {
-		label: 'Nothing', value: 'nothing'
-	}],
-	desktopNotificationDuration: [{
-		label: 'Default', value: 0
-	}, {
-		label: 'Seconds', second: 1, value: 1
-	}, {
-		label: 'Seconds', second: 2, value: 2
-	}, {
-		label: 'Seconds', second: 3, value: 3
-	}, {
-		label: 'Seconds', second: 4, value: 4
-	}, {
-		label: 'Seconds', second: 5, value: 5
-	}],
-	audioNotificationValue: [{
-		label: 'None', value: 'none None'
-	}, {
-		label: 'Default', value: '0 Default'
-	}, {
-		label: 'Beep', value: 'beep Beep'
-	}, {
-		label: 'Ding', value: 'ding Ding'
-	}, {
-		label: 'Chelle', value: 'chelle Chelle'
-	}, {
-		label: 'Droplet', value: 'droplet Droplet'
-	}, {
-		label: 'Highbell', value: 'highbell Highbell'
-	}, {
-		label: 'Seasons', value: 'seasons Seasons'
-	}]
-};
+import SectionTitle from './SectionTitle';
+import SectionSeparator from './SectionSeparator';
+import Info from './Info';
+import { OPTIONS } from './options';
 
 class NotificationPreferencesView extends React.Component {
 	static navigationOptions = () => ({
