+++ app/views/NotificationPreferencesView/options.js
@@ -1,0 +1,77 @@
+export const OPTIONS = {
+	desktopNotifications: [{
+		label: 'Default', value: 'default'
+	}, {
+		label: 'All_Messages', value: 'all'
+	}, {
+		label: 'Mentions', value: 'mentions'
+	}, {
+		label: 'Nothing', value: 'nothing'
+	}],
+	audioNotifications: [{
+		label: 'Default', value: 'default'
+	}, {
+		label: 'All_Messages', value: 'all'
+	}, {
+		label: 'Mentions', value: 'mentions'
+	}, {
+		label: 'Nothing', value: 'nothing'
+	}],
+	mobilePushNotifications: [{
+		label: 'Default', value: 'default'
+	}, {
+		label: 'All_Messages', value: 'all'
+	}, {
+		label: 'Mentions', value: 'mentions'
+	}, {
+		label: 'Nothing', value: 'nothing'
+	}],
+	mobileNotifications: [{
+		label: 'Default', value: 'default'
+	}, {
+		label: 'All_Messages', value: 'all'
+	}, {
+		label: 'Mentions', value: 'mentions'
+	}, {
+		label: 'Nothing', value: 'nothing'
+	}],
+	emailNotifications: [{
+		label: 'Default', value: 'default'
+	}, {
+		label: 'All_Messages', value: 'all'
+	}, {
+		label: 'Mentions', value: 'mentions'
+	}, {
+		label: 'Nothing', value: 'nothing'
+	}],
+	desktopNotificationDuration: [{
+		label: 'Default', value: 0
+	}, {
+		label: 'Seconds', second: 1, value: 1
+	}, {
+		label: 'Seconds', second: 2, value: 2
+	}, {
+		label: 'Seconds', second: 3, value: 3
+	}, {
+		label: 'Seconds', second: 4, value: 4
+	}, {
+		label: 'Seconds', second: 5, value: 5
+	}],
+	audioNotificationValue: [{
+		label: 'None', value: 'none None'
+	}, {
+		label: 'Default', value: '0 Default'
+	}, {
+		label: 'Beep', value: 'beep Beep'
+	}, {
+		label: 'Ding', value: 'ding Ding'
+	}, {
+		label: 'Chelle', value: 'chelle Chelle'
+	}, {
+		label: 'Droplet', value: 'droplet Droplet'
+	}, {
+		label: 'Highbell', value: 'highbell Highbell'
+	}, {
+		label: 'Seasons', value: 'seasons Seasons'
+	}]
+};
